fix(openapi-generator): fix more tests
